pyproject.toml
[tool.poetry]
name = "PJT_Automation"
version = "0.1.0"
description = ""
authors = ["mingchun1130 <74462405+mingchun1130@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "PJT_Automation"}]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.6.0"
webdriver-manager = "^3.8.4"
pytest = "^7.2.0"
pandas = "^1.5.1"
sqlalchemy = "^1.4.44"
pymysql = "^1.0.2"
allure-pytest = "^2.11.1"
python-dotenv = "^0.21.0"
openpyxl = "^3.0.10"
pytest-xdist = "^3.0.2"
pytest-rerunfailures = "^10.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
mypy = "^0.990"
isort = "^5.10.1"
types-requests = "^2.28.11.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vv -s"
testpaths = ["tests_web", "tests_api"]
pythonpath = "."
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "pytest.log"
log_file_level = "INFO"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"